change mc and arb to proxies
